Disable sampleShadingEnable for VulkanHelloAPI
The Vulkan specification states that "If [the sampleRateShading feature]
is not enabled, the sampleShadingEnable member of the
VkPipelineMultisampleStateCreateInfo structure must be set to VK_FALSE".

Since this app doesn't use sample rate shading, this can be disabled so
that it can run on Vulkan drivers that don't support this feature.